+++ ch_tools/chadmin/internal/object_storage/obj_list_item.py
@@ -1,0 +1,16 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class ObjListItem:
+    """
+    Item of object storage listing.
+    """
+
+    path: str
+    size: int
+
+    @classmethod
+    def from_tab_separated(cls, value: str) -> "ObjListItem":
+        path, size = value.split("\t")
+        return cls(path, int(size))
